requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
neo4j==5.14.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
Pillow==10.1.0
python-multipart==0.0.6
pydantic-settings==2.1.0